Update readMe.md
The most important aspect of this VM0042 methodology is missing from the description and calculations, which is the carbon stored in the soil SOC (soil organic carbon). So let's include this so that it can actually be used effectively. Remember that this methodology has nothing to do with REDD and should not be cited together with afforestation and reforestation projects. With the ongoing review of VM0042, some aspects of VM0047 will be included that will be used to calculate biomass. Let's follow along.

Signed-off-by: REGENA-BR <grove.bioconsult@gmail.com>